+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2]() - 2025-06-xx
+### Changed
+- Search template variable media_exclude to all_media. media_exclude will continue to work
+
+### Fixed
+- Enable thinking_budget for gemini-2.5-flash and gemini-2.5-pro only
+
 ## [1.0.1]() - 2025-06-13
 ### Fixed
 - Fix plex watch history id AttributeError; Fixes #10